Update sketch.js
there is a delay in the background image, first everything gets drawn and then the background image appears after some delay. So if we keep the entire code in if condition at line 50 so the delay is not there. 
So as a change I have applied brackets at line 51 and ended them at line 82.